requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
PyJWT==2.8.0
python-dotenv==1.0.0
flasgger==0.9.7.1
Werkzeug==2.3.8
pytest==7.4.3
Flask-Limiter
gunicorn==22.0.0
